coba2 check validation id

+++ components/products/repository.go
@@ -4,6 +4,7 @@ import (
 	"context"
 	"database/sql"
 	"errors"
+	"fmt"
 )
 
 func (d *Dependency) FindAll(ctx context.Context) ([]Product, error) {
@@ -48,49 +49,108 @@ func (d *Dependency) FindAll(ctx context.Context) ([]Product, error) {
 	return products, nil
 }
 
-func (d *Dependency) FindId(ctx context.Context, postId int) (*Product, error) {
+func (d *Dependency) FindId(ctx context.Context, postId int) (Product, error) {
 	conn, err := d.DB.Conn(ctx)
 	if err != nil {
-		return nil, errors.New("Connection failed!!!")
+		return Product{}, errors.New("Connection failed!!!")
+	}
+	defer conn.Close()
+
+	//check data exists or nor
+	row, err := d.DB.QueryContext(ctx,"SELECT EXISTS ( SELECT id FROM products WHERE id=?)",postId)
+	if err != nil {
+		return Product{},err
+	}
+	defer row.Close()
+	var exist bool
+	if row.Next(){
+		errScan := row.Scan(&exist)
+		if errScan != nil {
+			return Product{},err
+		}
+	}
+	if !exist {
+		return Product{}, errors.New("ID Was Not Found")
+	}
+
+	//if data is exists
+	row, err = d.DB.QueryContext(ctx, "SELECT id, name, stock, price FROM products WHERE id=?", postId)
+	if err != nil {
+		return Product{}, err
+	}
+	defer row.Close()
+	product := Product{}
+	if row.Next() {
+		err = row.Scan(&product.ID, &product.Name, &product.Stock, &product.Price)
+		if err != nil {
+			return Product{}, err
+		}
+	}
+	return product, nil
+}
+
+func (d *Dependency) Update2(ctx context.Context, postId int, product Product) (*Product, error) {
+	conn, err := d.DB.Conn(ctx)
+	if err != nil {
+		return nil, err
 	}
 	defer conn.Close()
 	tx, err := d.DB.BeginTx(ctx, &sql.TxOptions{Isolation: sql.LevelSerializable})
 	if err != nil {
 		return nil, err
 	}
+	//check ID
+	checkId, err := tx.QueryContext(ctx,"SELECT EXISTS ( SELECT id FROM products WHERE id=?)",postId)
+	if err != nil {
+		return nil, err
+	}
+	defer checkId.Close()
+	var exist bool
+	if checkId.Next(){
+		errScan := checkId.Scan(&exist)
+		if errScan != nil {
+			return nil, err
+		}
+	}
+	if !exist {
+		return nil, errors.New("ID Was Not Found")
+	}
 
-	row, execErr := tx.QueryContext(ctx, "SELECT id, name, stock, price FROM products WHERE id=?", postId)
+	//if check id was successfully and ID was found
+	_, execErr := tx.ExecContext(ctx, "UPDATE products SET name=?, stock=?, price=? WHERE id=?", &product.Name, &product.Stock, &product.Price, postId)
 	if execErr != nil {
 		err := tx.Rollback()
 		if err != nil {
 			return nil, err
 		}
-		return nil, execErr
+		return nil, errors.New("Execute error")
 	}
-	defer row.Close()
+	//row, errRow := res.RowsAffected()
+	//if errRow != nil {
+	//	errRoll := tx.Rollback()
+	//	if errRoll != nil {
+	//		return nil, errRoll
+	//	}
+	//	return nil, errRow
+	//}
+	//if row == 0 {
+	//	errRoll := tx.Rollback()
+	//	if errRoll != nil {
+	//		return nil, errRoll
+	//	}
+	//	return nil, nil
+	//}
 
-	product := &Product{}
-	if row.Next() {
-		err := row.Scan(&product.ID, &product.Name, &product.Stock, &product.Price)
-		if err != nil {
-			return nil, err
-		}
-		errCommit := tx.Commit()
-		if errCommit != nil {
-			err := tx.Rollback()
-			if err != nil {
-				return nil, err
-			}
-			return nil, errCommit
-		}
-		return product, nil
-	} else {
+	product.ID = postId
+	errCommit := tx.Commit()
+	if errCommit != nil {
 		err := tx.Rollback()
 		if err != nil {
 			return nil, err
 		}
-		return nil, err
+		return nil, errCommit
 	}
+	return &product, nil
 }
 
 func (d *Dependency) Update(ctx context.Context, postId int, product Product) (*Product, error) {
@@ -103,6 +163,22 @@ func (d *Dependency) Update(ctx context.Context, postId int, product Product) (*
 	if err != nil {
 		return nil, err
 	}
+
+	rows, err := tx.QueryContext(ctx,"SELECT EXISTS ( SELECT id FROM products WHERE id=?)",postId)
+	if err != nil {
+		return nil, errors.New("err query")
+	}
+	var exists bool
+	if rows.Next() {
+		err = rows.Scan(&exists)
+		if err!=nil {
+			return nil, err
+		}
+	}
+	if !exists {
+		return nil, errors.New("not found")
+	}
+
 	res, execErr := tx.ExecContext(ctx, "UPDATE products SET name=?, stock=?, price=? WHERE id=?", &product.Name, &product.Stock, &product.Price, postId)
 	if execErr != nil {
 		err := tx.Rollback()
@@ -139,7 +215,7 @@ func (d *Dependency) Update(ctx context.Context, postId int, product Product) (*
 	return &product, nil
 }
 
-func (d *Dependency) Delete(ctx context.Context, postId int) error {
+func (d *Dependency) Delete2(ctx context.Context, postId int) error {
 	conn, err := d.DB.Conn(ctx)
 	if err != nil {
 		return errors.New("Connection failed!!!")
@@ -149,22 +225,38 @@ func (d *Dependency) Delete(ctx context.Context, postId int) error {
 	if err != nil {
 		return err
 	}
+	row, err := tx.QueryContext(ctx,"SELECT EXISTS ( SELECT id FROM products WHERE id=?)",postId)
+	if err != nil {
+		return err
+	}
+	defer row.Close()
+	var exist bool
+	if row.Next(){
+		errScan := row.Scan(&exist)
+		if errScan != nil {
+			return err
+		}
+	}
+	if !exist {
+		return errors.New("ID Was Not Found")
+	}
 
+	//if check id was successfully and ID was found
 	_, err = tx.ExecContext(ctx,"DELETE FROM products WHERE id =? ",postId)
 	if err != nil {
 		errRoll := tx.Rollback()
 		if errRoll != nil {
 			return errRoll
 		}
-		return err
+		return errors.New("exec error")
 	}
 	errCommit := tx.Commit()
 	if errCommit != nil {
 		err := tx.Rollback()
 		if err != nil {
-			return err
+			return errors.New("roll error")
 		}
-		return errCommit
+		return errors.New("Commit error")
 	}
 	return nil
 }
@@ -206,4 +298,52 @@ func (d *Dependency)Save(ctx context.Context, product Product) (*Product, error)
 	}
 	product.ID = int(id)
 	return &product, nil
+}
+
+func (d *Dependency) Delete(ctx context.Context, postId int) error {
+	conn, err := d.DB.Conn(ctx)
+	if err != nil {
+		return errors.New("Connection failed!!!")
+	}
+	defer conn.Close()
+	tx, err := conn.BeginTx(ctx, nil)
+	if err != nil {
+		return err
+	}
+
+	row, err := tx.QueryContext(ctx,"SELECT EXISTS ( SELECT * FROM products WHERE id=?)",postId)
+	if err != nil {
+		return err
+	}
+	defer row.Close()
+	var exist bool
+	if row.Next(){
+		errScan := row.Scan(&exist)
+		if errScan != nil {
+			return err
+		}
+	}
+	if !exist {
+		return errors.New("ID Was Not Found")
+	}
+
+	//if check id was successfully and ID was found
+	_, err = tx.ExecContext(ctx,"DELETE FROM products WHERE id =? ",postId)
+	if err != nil {
+		errRoll := tx.Rollback()
+		if errRoll != nil {
+			fmt.Println(errRoll)
+			return errors.New("rollback error")
+		}
+		return errors.New("exec error")
+	}
+	errCommit := tx.Commit()
+	if errCommit != nil {
+		err := tx.Rollback()
+		if err != nil {
+			return errors.New("roll error")
+		}
+		return errors.New("Commit error")
+	}
+	return nil
 }
